+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - added fade animations
 - fixed certain inputs skipping too many things
 - fixed a memory leak in the audio sampler
+- added a vsync option
 
 
 ## [2.4](https://github.com/rr-/Tomb1Main/compare/2.3...2.4) - 2022-01-19